extract axes rendering from graphLayer
