+++ demo/src/components/demos/graph.js
@@ -108,8 +108,9 @@ export default class GraphDemo extends Component {
       xRange: [-Math.PI, Math.PI],
       yRange: [-Math.PI, Math.PI],
       resolution: [resolution.value, resolution.value],
-      axisOffset: 0.5,
-      axisColor: showAxis.value ? [0, 0, 0, 128] : [0, 0, 0, 0],
+      drawAxes: showAxis.value,
+      axesOffset: 0.5,
+      axesColor: [0, 0, 0, 128],
       opacity: 1,
       pickable: true,
       onHover: this._onHover,
